Search thesis, lecturer, student and fix bugs

+++ api/src/lecturer/lecturer.service.ts
@@ -20,17 +20,22 @@ export class LecturerService {
     private readonly connection: Connection
   ) {}
 
-  public async findMany(offset: number, limit: number): Promise<Lecturer[]> {
+  public async getMany(offset: number, limit: number, keyword?: string): Promise<Lecturer[]> {
+    let conditions: FindOptionsWhere<Lecturer> | undefined = undefined;
+    if (keyword) {
+      conditions = this.getSearchConditions(keyword);
+    }
+
     return await this.lecturerRepository.find({
-      relations: [LecturerRelation.USER],
-      where: { ...NOT_DELETE_CONDITION },
+      relations: { user: {} },
+      where: conditions ? conditions : { ...NOT_DELETE_CONDITION },
       skip: offset,
       take: limit,
       cache: true
     });
   }
 
-  public async findById(id: number): Promise<Lecturer> {
+  public async getById(id: number): Promise<Lecturer> {
     const lecturer: Lecturer | undefined = await this.lecturerRepository.findOne(id, {
       relations: [LecturerRelation.USER],
       where: { ...NOT_DELETE_CONDITION }
@@ -92,8 +97,15 @@ export class LecturerService {
       return;
     }
 
-    const currentLecturer = await this.findById(id);
+    const currentLecturer = await this.getById(id);
     if (userBody) {
+      if (
+        userBody.status === UserStatus.INACTIVE &&
+        (await this.thesisLecturerService.isLecturerParticipatedThesis(id))
+      ) {
+        throw new BadRequestException(LecturerError.ERR_4);
+      }
+
       currentLecturer.user = await this.userService.updateById(currentLecturer.user, userBody);
     }
 
@@ -117,8 +129,11 @@ export class LecturerService {
   }
 
   public async deleteById(id: number): Promise<void> {
-    const lecturer = await this.findById(id);
+    const lecturer = await this.getById(id);
     const deletedAt = new Date();
+    if (await this.thesisLecturerService.isLecturerParticipatedThesis(id)) {
+      throw new BadRequestException(LecturerError.ERR_4);
+    }
 
     await this.connection.transaction(async (manager) => {
       await this.thesisLecturerService.deleteByLecturerIdWithTransaction(manager, id, deletedAt);
@@ -128,8 +143,13 @@ export class LecturerService {
     });
   }
 
-  public async getLecturerAmount(): Promise<number> {
-    return this.lecturerRepository.count({ ...NOT_DELETE_CONDITION });
+  public async getLecturerAmount(keyword?: string): Promise<number> {
+    let conditions: FindOptionsWhere<Lecturer> | undefined = undefined;
+    if (keyword) {
+      conditions = this.getSearchConditions(keyword);
+    }
+
+    return this.lecturerRepository.count(conditions ? conditions : { ...NOT_DELETE_CONDITION });
   }
 
   public sanitizeLevel(level: string): string {
@@ -224,5 +244,36 @@ export class LecturerService {
         UserError.ERR_9.replace('%s', this.generateErrorInfo(lecturer))
       );
     }
+  }
+
+  private getSearchConditions(keyword: string): FindOptionsWhere<Lecturer> {
+    return [
+      {
+        ...NOT_DELETE_CONDITION,
+        lecturerId: Like(`%${keyword}%`)
+      },
+      {
+        ...NOT_DELETE_CONDITION,
+        position: Like(`%${keyword}%`)
+      },
+      {
+        user: {
+          ...NOT_DELETE_CONDITION,
+          username: Like(`%${keyword}%`)
+        }
+      },
+      {
+        user: {
+          ...NOT_DELETE_CONDITION,
+          firstname: Like(`%${keyword}%`)
+        }
+      },
+      {
+        user: {
+          ...NOT_DELETE_CONDITION,
+          lastname: Like(`%${keyword}%`)
+        }
+      }
+    ];
   }
 }
